- also relocate non-USJ nodes
