+++ packages/shared/nodes/features/ImmutableUnmatchedNode.ts
@@ -11,7 +11,7 @@ import {
   isHTMLElement,
   DOMConversionOutput,
 } from "lexical";
-import { INVALID_CLASS_NAME, ZWSP } from "./node.utils";
+import { INVALID_CLASS_NAME, ZWSP } from "../scripture/usj/node.utils";
 
 export const UNMATCHED_TAG_NAME = "unmatched";
 export const IMMUTABLE_UNMATCHED_VERSION = 1;
